app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
boto3==1.40.26
botocore==1.40.26
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.1
Django==5.1.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-jalali==7.4.0
django-redis==5.4.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-prometheus>=2.3.0
drf-spectacular==0.28.0
flower==2.0.1
gunicorn==23.0.0
hiredis==3.2.1
humanize==4.13.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jalali_core==1.0.0
jdatetime==5.2.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kavenegar==1.1.2
Khayyam==3.0.17
kombu==5.4.2
# mysql-connector-python==9.2.0
packaging==24.2
pika==1.3.2
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
